[PowerPC] Add dcbz_l extension instruction
The Locked cache is a feature implemented by the PowerPC Gekko, 750CL, Broadway and
Espresso CPUs. It allows the fixing of specific  cache lines in L2 cache
for DMA. An instruction, dcbz_l, was added to accomodate this extension.
It zeroes the selected line and locks it into the cache

Signed-off-by: Charlotte Delenk <darkkirb@darkkirb.de>